+++ saleor/graphql/account/mutations/base.py
@@ -128,7 +128,12 @@ class RequestPasswordReset(BaseMutation):
     def perform_mutation(cls, _root, info, **data):
         email = data["email"]
         redirect_url = data["redirect_url"]
-        validate_storefront_url(redirect_url)
+        try:
+            validate_storefront_url(redirect_url)
+        except ValidationError as error:
+            raise ValidationError(
+                {"redirect_url": error}, code=AccountErrorCode.INVALID
+            )
 
         try:
             user = models.User.objects.get(email=email)
@@ -297,7 +302,11 @@ class CustomerInput(UserInput, UserAddressInput):
 
 class UserCreateInput(CustomerInput):
     send_password_email = graphene.Boolean(
-        description="Send an email with a link to set a password."
+        description=(
+            "DEPRECATED: Will be removed in Saleor 2.10, if mutation has `redirect_url`"
+            " in input then customer get email with link to set a password. "
+            "Send an email with a link to set a password."
+        )
     )
     redirect_url = graphene.String(
         description=(
@@ -337,6 +346,8 @@ class BaseCustomerCreate(ModelMutation, I18nMixin):
             )
             cleaned_input[BILLING_ADDRESS_FIELD] = billing_address
 
+        # DEPRECATED: We should remove this condition when dropping
+        # `send_password_email` from mutation input.
         if cleaned_input.get("send_password_email"):
             if not cleaned_input.get("redirect_url"):
                 raise ValidationError(
@@ -347,7 +358,14 @@ class BaseCustomerCreate(ModelMutation, I18nMixin):
                         )
                     }
                 )
-            validate_storefront_url(cleaned_input.get("redirect_url"))
+
+        if cleaned_input.get("redirect_url"):
+            try:
+                validate_storefront_url(cleaned_input.get("redirect_url"))
+            except ValidationError as error:
+                raise ValidationError(
+                    {"redirect_url": error}, code=AccountErrorCode.INVALID
+                )
 
         return cleaned_input
 
@@ -372,7 +390,7 @@ class BaseCustomerCreate(ModelMutation, I18nMixin):
             info.context.extensions.customer_created(customer=instance)
             account_events.customer_account_created_event(user=instance)
 
-        if cleaned_input.get("send_password_email"):
+        if cleaned_input.get("redirect_url"):
             send_set_password_email_with_url(
                 cleaned_input.get("redirect_url"), instance
             )