+++ saleor/site/models.py
@@ -61,7 +61,6 @@ class SiteSettings(models.Model):
     company_address = models.ForeignKey(
         "account.Address", blank=True, null=True, on_delete=models.CASCADE
     )
-
     default_mail_sender_name = models.CharField(
         max_length=settings.DEFAULT_MAX_EMAIL_DISPLAY_NAME_LENGTH,
         blank=True,
@@ -69,7 +68,7 @@ class SiteSettings(models.Model):
         validators=EMAIL_SENDER_NAME_VALIDATORS,
     )
     default_mail_sender_address = models.EmailField(blank=True, null=True)
-
+    customer_set_password_url = models.CharField(max_length=255, blank=True, null=True)
     translated = TranslationProxy()
 
     class Meta: