+++ tests/api/test_account.py
@@ -646,6 +646,10 @@ CUSTOMER_CREATE_MUTATION = """
                 field
                 message
             }
+            accountErrors {
+                field
+                code
+            }
             user {
                 id
                 defaultBillingAddress {
@@ -683,7 +687,6 @@ def test_customer_create(
         "note": note,
         "shipping": address_data,
         "billing": address_data,
-        "send_mail": True,
         "redirect_url": "https://www.example.com",
     }
 
@@ -727,11 +730,7 @@ def test_customer_create_send_password_with_url(
     permission_manage_users,
 ):
     email = "api_user@example.com"
-    variables = {
-        "email": email,
-        "send_mail": True,
-        "redirect_url": "https://www.example.com",
-    }
+    variables = {"email": email, "redirect_url": "https://www.example.com"}
 
     response = staff_api_client.post_graphql(
         CUSTOMER_CREATE_MUTATION, variables, permissions=[permission_manage_users]
@@ -763,7 +762,7 @@ def test_customer_create_without_send_password(
     User.objects.get(email=email)
 
 
-def test_customer_create_without_redirect_url(
+def test_customer_create_without_redirect_url_deprecated_send_mail_flag(
     staff_api_client, permission_manage_users
 ):
     email = "api_user@example.com"
@@ -773,20 +772,26 @@ def test_customer_create_without_redirect_url(
     )
     content = get_graphql_content(response)
     data = content["data"]["customerCreate"]
-    assert data["errors"][0]["field"] == "redirectUrl"
+    assert data["accountErrors"][0] == {
+        "field": "redirectUrl",
+        "code": AccountErrorCode.REQUIRED.name,
+    }
     staff_user = User.objects.filter(email=email)
     assert not staff_user
 
 
 def test_customer_create_with_invalid_url(staff_api_client, permission_manage_users):
     email = "api_user@example.com"
-    variables = {"email": email, "send_mail": True, "redirect_url": "invalid"}
+    variables = {"email": email, "redirect_url": "invalid"}
     response = staff_api_client.post_graphql(
         CUSTOMER_CREATE_MUTATION, variables, permissions=[permission_manage_users]
     )
     content = get_graphql_content(response)
     data = content["data"]["customerCreate"]
-    assert data["errors"][0]["field"] == "redirectUrl"
+    assert data["accountErrors"][0] == {
+        "field": "redirectUrl",
+        "code": AccountErrorCode.INVALID.name,
+    }
     staff_user = User.objects.filter(email=email)
     assert not staff_user
 
@@ -795,17 +800,16 @@ def test_customer_create_with_not_allowed_url(
     staff_api_client, permission_manage_users
 ):
     email = "api_user@example.com"
-    variables = {
-        "email": email,
-        "send_mail": True,
-        "redirectUrl": "https://www.fake.com",
-    }
+    variables = {"email": email, "redirect_url": "https://www.fake.com"}
     response = staff_api_client.post_graphql(
         CUSTOMER_CREATE_MUTATION, variables, permissions=[permission_manage_users]
     )
     content = get_graphql_content(response)
     data = content["data"]["customerCreate"]
-    assert data["errors"][0]["field"] == "redirectUrl"
+    assert data["accountErrors"][0] == {
+        "field": "redirectUrl",
+        "code": AccountErrorCode.INVALID.name,
+    }
     staff_user = User.objects.filter(email=email)
     assert not staff_user
 
@@ -1110,6 +1114,10 @@ ACCOUNT_REQUEST_DELETION_MUTATION = """
                 field
                 message
             }
+            accountErrors {
+                code
+                field
+            }
         }
     }
 """
@@ -1173,7 +1181,10 @@ def test_account_request_deletion_storefront_hosts_not_allowed(
     content = get_graphql_content(response)
     data = content["data"]["accountRequestDeletion"]
     assert len(data["errors"]) == 1
-    assert data["errors"][0]["field"] == "redirectUrl"
+    assert data["accountErrors"][0] == {
+        "field": "redirectUrl",
+        "code": AccountErrorCode.INVALID.name,
+    }
     send_account_delete_confirmation_email_with_url_mock.assert_not_called()
 
 
@@ -1358,7 +1369,6 @@ STAFF_CREATE_MUTATION = """
             }
             accountErrors {
                 field
-                message
                 code
             }
             user {
@@ -1386,7 +1396,6 @@ def test_staff_create(
     variables = {
         "email": email,
         "permissions": [PermissionEnum.MANAGE_PRODUCTS.name],
-        "send_mail": True,
         "redirect_url": "https://www.example.com",
     }
 
@@ -1423,11 +1432,7 @@ def test_staff_create_send_password_with_url(
     permission_manage_staff,
 ):
     email = "api_user@example.com"
-    variables = {
-        "email": email,
-        "send_mail": True,
-        "redirect_url": "https://www.example.com",
-    }
+    variables = {"email": email, "redirect_url": "https://www.example.com"}
 
     response = staff_api_client.post_graphql(
         STAFF_CREATE_MUTATION, variables, permissions=[permission_manage_staff]
@@ -1459,7 +1464,7 @@ def test_staff_create_without_send_password(
     User.objects.get(email=email)
 
 
-def test_staff_create_without_redirect_url(
+def test_staff_create_without_redirect_url_deprecated_send_mail_flag(
     staff_api_client, media_root, permission_manage_staff
 ):
     email = "api_user@example.com"
@@ -1469,7 +1474,10 @@ def test_staff_create_without_redirect_url(
     )
     content = get_graphql_content(response)
     data = content["data"]["staffCreate"]
-    assert data["errors"][0]["field"] == "redirectUrl"
+    assert data["accountErrors"][0] == {
+        "field": "redirectUrl",
+        "code": AccountErrorCode.REQUIRED.name,
+    }
     staff_user = User.objects.filter(email=email)
     assert not staff_user
 
@@ -1478,13 +1486,16 @@ def test_staff_create_with_invalid_url(
     staff_api_client, media_root, permission_manage_staff
 ):
     email = "api_user@example.com"
-    variables = {"email": email, "send_mail": True, "redirect_url": "invalid"}
+    variables = {"email": email, "redirect_url": "invalid"}
     response = staff_api_client.post_graphql(
         STAFF_CREATE_MUTATION, variables, permissions=[permission_manage_staff]
     )
     content = get_graphql_content(response)
     data = content["data"]["staffCreate"]
-    assert data["errors"][0]["field"] == "redirectUrl"
+    assert data["accountErrors"][0] == {
+        "field": "redirectUrl",
+        "code": AccountErrorCode.INVALID.name,
+    }
     staff_user = User.objects.filter(email=email)
     assert not staff_user
 
@@ -1492,18 +1503,17 @@ def test_staff_create_with_invalid_url(
 def test_staff_create_with_not_allowed_url(
     staff_api_client, media_root, permission_manage_staff
 ):
-    email = "api_user@example.com"
-    variables = {
-        "email": email,
-        "send_mail": True,
-        "redirectUrl": "https://www.fake.com",
-    }
+    email = "api_userrr@example.com"
+    variables = {"email": email, "redirect_url": "https://www.fake.com"}
     response = staff_api_client.post_graphql(
         STAFF_CREATE_MUTATION, variables, permissions=[permission_manage_staff]
     )
     content = get_graphql_content(response)
     data = content["data"]["staffCreate"]
-    assert data["errors"][0]["field"] == "redirectUrl"
+    assert data["accountErrors"][0] == {
+        "field": "redirectUrl",
+        "code": AccountErrorCode.INVALID.name,
+    }
     staff_user = User.objects.filter(email=email)
     assert not staff_user
 
@@ -1558,7 +1568,6 @@ def test_staff_update_doesnt_change_existing_avatar(
             }
             accountErrors {
                 field
-                message
                 code
             }
         }